+++ imports/pages/devotionals/__tests__/Single.js
@@ -96,19 +96,13 @@ it("renders with scripture", () => {
 it("dispatches to the store on mount", () => {
   const mockDispatch = jest.fn();
   liveActions.hide = jest.fn();
-  navActions.setLevel = jest.fn();
-  navActions.setAction = jest.fn();
   headerActions.set = jest.fn();
   headerActions.hide = jest.fn();
   const wrapper = shallow(generateComponent({
     dispatch: mockDispatch,
   }));
-  expect(mockDispatch).toHaveBeenCalledTimes(5);
+  expect(mockDispatch).toHaveBeenCalledTimes(3);
   expect(liveActions.hide).toHaveBeenCalledTimes(1);
-  expect(navActions.setLevel).toHaveBeenCalledTimes(1);
-  expect(navActions.setLevel).toHaveBeenCalledWith("CONTENT");
-  expect(navActions.setAction).toHaveBeenCalledTimes(1);
-  expect(navActions.setAction.mock.calls[0][0]).toBe("CONTENT");
   expect(headerActions.set).toHaveBeenCalledTimes(1);
   expect(headerActions.hide).toHaveBeenCalledTimes(1);
 });
@@ -159,7 +153,7 @@ it("unfloats the live bar on unmount", () => {
     dispatch: mockDispatch,
   }));
   wrapper.instance().componentWillUnmount();
-  expect(mockDispatch).toHaveBeenCalledTimes(6);
+  expect(mockDispatch).toHaveBeenCalledTimes(4);
   expect(liveActions.unfloat).toHaveBeenCalledTimes(1);
 });