+++ imports/pages/events/index.js
@@ -1,0 +1,13 @@
+// @flow
+import Single from "./Single";
+
+const Event = () => <Single />;
+
+const Routes = [
+  { path: "events/:id", component: Single },
+];
+
+export default {
+  Event,
+  Routes,
+};